api automation with http clients

+++ src/main/java/warmUpQuestions/FindAsciiValue.java
@@ -3,8 +3,8 @@ package warmUpQuestions;
 public class FindAsciiValue {
     public static void main(String[] args) {
         //Now, to find the ASCII value of ch, we just assign ch to an int variable ascii. Internally, Java converts the character value to an ASCII value.
-        char a='a';
-        int asci=(int) a;
+        char a = 'a';
+        int asci = a;
         System.out.println(asci);
     }
 }
